[RPI] improving fail proof for assignment : even if the front fails to prevent a bad assignment, the assignment can't be partially done

+++ manifmaker/both/methods/meteor-method.js
@@ -79,19 +79,21 @@ Meteor.methods({
         }
 
 
-        var assignmentId = Assignments.insert({
-            userId: userId,
-            taskId: task._id,
-            timeSlotId: timeSlot._id,
-            peopleNeedId: peopleNeed._id
-        });
+        if(PeopleNeedService.assignedPeopleNeeded(task, timeSlot, peopleNeed, userId) === 1) {
 
+            var assignmentId = Assignments.insert({
+                userId: userId,
+                taskId: task._id,
+                timeSlotId: timeSlot._id,
+                peopleNeedId: peopleNeed._id
+            });
 
-        AvailabilityService.removeAvailabilities(user, timeSlot.start, timeSlot.end);
-        PeopleNeedService.assignedPeopleNeeded(task, timeSlot, peopleNeed, userId);
+            AvailabilityService.removeAvailabilities(user, timeSlot.start, timeSlot.end);
 
+            return assignmentId;
+        }
 
-        return assignmentId;
+        return null;
     }
 });
